Fix issue when CSV data does not end with a newline
If CSV data didn't end with "\n", "\r", or "\r\n" the last line of the
CSV would be ignored, as the parser wouldn't trigger end of row logic
for the last row.

This change forces the parser to process the end of the CSV data as it's
own character (false), and deals with end of data just like it treats
end of row.